+++ Amplify/Categories/DataStore/Model/Temporal/AnyTemporalSpec.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 /// Type-erased protocol for `TemporalSpec` providing access to `iso8601String`
-public protocol _AnyTemporalSpec {
+public protocol AnyTemporalSpec {
     /// The ISO-8601 formatted string in the UTC `TimeZone`.
     var iso8601String: String { get }
 }
