+++ Amplify/Categories/DataStore/Model/Temporal/DateUnitOperatable.swift
@@ -8,12 +8,12 @@
 import Foundation
 
 /// Supports addition and subtraction of `Temporal.Date` and `Temporal.DateTime` with `DateUnit`
-public protocol _DateUnitOperable {
-    static func + (left: Self, right: _DateUnit) -> Self
-    static func - (left: Self, right: _DateUnit) -> Self
+public protocol DateUnitOperable {
+    static func + (left: Self, right: DateUnit) -> Self
+    static func - (left: Self, right: DateUnit) -> Self
 }
 
-extension _TemporalSpec where Self: _DateUnitOperable {
+extension TemporalSpec where Self: DateUnitOperable {
 
     /// Add a `DateUnit` to a `Temporal.Date` or `Temporal.DateTime`
     ///
@@ -23,7 +23,7 @@ extension _TemporalSpec where Self: _DateUnitOperable {
     ///   - left: `Temporal.Date` or `Temporal.DateTime`
     ///   - right: `DateUnit` to add to `left`
     /// - Returns: A new `Temporal.Date` or `Temporal.DateTime` the `DateUnit` was added to.
-    public static func + (left: Self, right: _DateUnit) -> Self {
+    public static func + (left: Self, right: DateUnit) -> Self {
         return left.add(value: right.value, to: right.calendarComponent)
     }
 
@@ -35,7 +35,7 @@ extension _TemporalSpec where Self: _DateUnitOperable {
     ///   - left: `Temporal.Date` or `Temporal.DateTime`
     ///   - right: `DateUnit` to subtract from `left`
     /// - Returns: A new `Temporal.Date` or `Temporal.DateTime` the `DateUnit` was subtracted from.
-    public static func - (left: Self, right: _DateUnit) -> Self {
+    public static func - (left: Self, right: DateUnit) -> Self {
         return left.add(value: -right.value, to: right.calendarComponent)
     }
 }
