+++ Amplify/Categories/DataStore/Model/Temporal/Temporal+Codable.swift
@@ -5,27 +5,9 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
+import Foundation
+
 extension TemporalSpec where Self: Codable {
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        let value = try container.decode(String.self)
-        try self.init(iso8601String: value)
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-        try container.encode(iso8601String)
-    }
-
-}
-
-extension Temporal.Date: Codable {}
-extension Temporal.DateTime: Codable {}
-extension Temporal.Time: Codable {}
-
-extension _TemporalSpec where Self: Codable {
-
     public init(from decoder: Decoder) throws {
         let container = try decoder.singleValueContainer()
         let value = try container.decode(String.self)
@@ -36,9 +18,8 @@ extension _TemporalSpec where Self: Codable {
         var container = encoder.singleValueContainer()
         try container.encode(iso8601String)
     }
-
 }
 
-extension _Temporal.Date: Codable {}
-extension _Temporal.DateTime: Codable {}
-extension _Temporal.Time: Codable {}
+extension Temporal.Date: Codable {}
+extension Temporal.DateTime: Codable {}
+extension Temporal.Time: Codable {}
