+++ Amplify/Categories/DataStore/Model/Temporal/Temporal+Hashable.swift
@@ -8,9 +8,7 @@
 import Foundation
 
 extension TemporalSpec where Self: Hashable {
-
     func hash(into hasher: inout Hasher) {
         hasher.combine(foundationDate)
     }
-
 }
