+++ Amplify/Categories/DataStore/Model/Temporal/TimeUnitOperatable.swift
@@ -8,19 +8,19 @@
 import Foundation
 
 /// Supports addition and subtraction of `Temporal.Time` and `Temporal.DateTime` with `TimeUnit`
-public protocol _TimeUnitOperable {
-    static func + (left: Self, right: _TimeUnit) -> Self
-    static func - (left: Self, right: _TimeUnit) -> Self
+public protocol TimeUnitOperable {
+    static func + (left: Self, right: TimeUnit) -> Self
+    static func - (left: Self, right: TimeUnit) -> Self
 }
 
-extension _TemporalSpec where Self: _TimeUnitOperable {
+extension TemporalSpec where Self: TimeUnitOperable {
 
     /// Add a `TimeUnit` to a `Temporal.Time` or `Temporal.DateTime`
     /// - Parameters:
     ///   - left: `Temporal.Time` or `Temporal.DateTime`
     ///   - right: `TimeUnit` to add to `left`
     /// - Returns: A new `Temporal.Time` or `Temporal.DateTime` the `TimeUnit` was added to.
-    public static func + (left: Self, right: _TimeUnit) -> Self {
+    public static func + (left: Self, right: TimeUnit) -> Self {
         return left.add(value: right.value, to: right.calendarComponent)
     }
 
@@ -29,7 +29,7 @@ extension _TemporalSpec where Self: _TimeUnitOperable {
     ///   - left: `Temporal.Time` or `Temporal.DateTime`
     ///   - right: `TimeUnit` to subtract from `left`
     /// - Returns: A new `Temporal.Time` or `Temporal.DateTime` the `TimeUnit` was subtracted from.
-    public static func - (left: Self, right: _TimeUnit) -> Self {
+    public static func - (left: Self, right: TimeUnit) -> Self {
         return left.add(value: -right.value, to: right.calendarComponent)
     }
 }
